apply changes from #1762 for #1749 manually

+++ lib/Migration/DeleteInvalidRecords.php
@@ -1,8 +1,7 @@
 <?php
 /**
- * @copyright Copyright (c) 2017 Julius Härtl <jus@bitgrid.net>
+ * @copyright Copyright (c) 2021 René Gieling <github@dartcafe.de>
  *
- * @author Julius Härtl <jus@bitgrid.net>
  * @author René Gieling <github@dartcafe.de>
  *
  * @license GNU AGPL version 3 or any later version
@@ -26,6 +25,7 @@
 namespace OCA\Polls\Migration;
 
 use OC\DB\Connection;
+use OC\DB\SchemaWrapper;
 use OCP\IConfig;
 use OCP\Migration\IRepairStep;
 use OCP\Migration\IOutput;
@@ -100,13 +100,16 @@ class DeleteInvalidRecords implements IRepairStep {
 	}
 
 	public function run(IOutput $output):void {
-		$this->removeOrphaned();
-		$this->logMapper->removeDuplicates($output);
-		$this->optionMapper->removeDuplicates($output);
-		$this->preferencesMapper->removeDuplicates($output);
-		$this->shareMapper->removeDuplicates($output);
-		$this->subscriptionMapper->removeDuplicates($output);
-		$this->voteMapper->removeDuplicates($output);
+		$schema = new SchemaWrapper($this->connection);
+		if ($schema->hasTable('polls_polls')) {
+			$this->removeOrphaned();
+			$this->logMapper->removeDuplicates($output);
+			$this->optionMapper->removeDuplicates($output);
+			$this->preferencesMapper->removeDuplicates($output);
+			$this->shareMapper->removeDuplicates($output);
+			$this->subscriptionMapper->removeDuplicates($output);
+			$this->voteMapper->removeDuplicates($output);
+		}
 	}
 
 	/**
